pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.12.1",
    "google-genai>=1.56.0",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]
